+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   - increase test coverage
   - check for Python minimal version 3.6
+  - new endpoint `v3/connect/loginFromSession`
 
 
 ## [20191112]